requirements.txt
PyTorch for CUDA 11.8
torch==2.1.0 --index-url https://download.pytorch.org/whl/cu118
torchaudio==2.1.0 --index-url https://download.pytorch.org/whl/cu118
torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cu118

RVC Core Dependencies
numpy==1.23.5
librosa==0.9.2
numba==0.56.4
fairseq==0.12.2
gradio==3.14.0
pyworld
praat-parselmouth
ffmpeg-python